Adds function to add recipe ingredients
When adding a new list, you can now choose if you want to add a new shopping list or a new chefkoch(recipe) list. When choosing recipe then you can enter a link or the id to the chefkoch recipe. Then a new shopping list with all ingredients will be added.